Structuralized AppDelegate and fixed token flow.
