+++ QuickStart/AppDelegate+PushKit.swift
@@ -1,0 +1,62 @@
+//
+//  AppDelegate+PushKit.swift
+//  QuickStart
+//
+//  Created by Minhyuk Kim on 2020/05/11.
+//  Copyright © 2020 SendBird Inc. All rights reserved.
+//
+
+import Foundation
+import SendBirdCalls
+import CallKit
+import PushKit
+
+// MARK: - PushKit
+extension AppDelegate: PKPushRegistryDelegate {
+    func voipRegistration() {
+        self.voipRegistry = PKPushRegistry(queue: DispatchQueue.main)
+        self.voipRegistry?.delegate = self
+        self.voipRegistry?.desiredPushTypes = [.voIP]
+    }
+    
+    func pushRegistry(_ registry: PKPushRegistry, didInvalidatePushTokenFor type: PKPushType) {
+        SendBirdCall.unregisterVoIPPush(token: UserDefaults.standard.voipPushToken) { error in
+            guard error == nil else { return }
+            UserDefaults.standard.voipPushToken = nil
+        }
+    }
+    
+    // MARK: SendBirdCalls - Registering push token.
+    func pushRegistry(_ registry: PKPushRegistry, didUpdate pushCredentials: PKPushCredentials, for type: PKPushType) {
+        UserDefaults.standard.voipPushToken = pushCredentials.token
+        print("Push token is \(pushCredentials.token.toHexString())")
+        
+        SendBirdCall.registerVoIPPush(token: pushCredentials.token, unique: true) { error in
+            guard error == nil else { return }
+        }
+    }
+    
+    // MARK: SendBirdCalls - Receive incoming push event
+    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType) {
+        SendBirdCall.pushRegistry(registry, didReceiveIncomingPushWith: payload, for: type, completionHandler: nil)
+    }
+    
+    // Handle Incoming pushes
+    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType, completion: @escaping () -> Void) {
+        // MARK: Handling incoming call
+        SendBirdCall.pushRegistry(registry, didReceiveIncomingPushWith: payload, for: type) { uuid in
+            guard uuid != nil else {
+                let update = CXCallUpdate()
+                update.remoteHandle = CXHandle(type: .generic, value: "invalid")
+                let randomUUID = UUID()
+                CXCallManager.shared.reportIncomingCall(with: randomUUID, update: update) { error in
+                    CXCallManager.shared.endCall(for: randomUUID, endedAt: Date(), reason: .unknown)
+                }
+                completion()
+                return
+            }
+
+            completion()
+        }
+    }
+}
