+++ QuickStart/AppDelegate.swift
@@ -69,89 +69,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // In this case, I recommend that you register local notification to notify the unterminated calls.
     }
 }
-
-// MARK: - PushKit
-extension AppDelegate: PKPushRegistryDelegate {
-    func voipRegistration() {
-        self.voipRegistry = PKPushRegistry(queue: DispatchQueue.main)
-        self.voipRegistry?.delegate = self
-        self.voipRegistry?.desiredPushTypes = [.voIP]
-    }
-    
-    func pushRegistry(_ registry: PKPushRegistry, didInvalidatePushTokenFor type: PKPushType) {
-        SendBirdCall.unregisterVoIPPush(token: UserDefaults.standard.voipPushToken) { error in
-            guard error == nil else { return }
-        }
-        UserDefaults.standard.voipPushToken = nil
-    }
-    
-    // MARK: SendBirdCalls - Registering push token.
-    func pushRegistry(_ registry: PKPushRegistry, didUpdate pushCredentials: PKPushCredentials, for type: PKPushType) {
-        UserDefaults.standard.voipPushToken = pushCredentials.token
-        print("Push token is \(pushCredentials.token.toHexString())")
-        
-        SendBirdCall.registerVoIPPush(token: pushCredentials.token, unique: true) { error in
-            guard error == nil else { return }
-        }
-    }
-    
-    // MARK: SendBirdCalls - Receive incoming push event
-    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType) {
-        SendBirdCall.pushRegistry(registry, didReceiveIncomingPushWith: payload, for: type, completionHandler: nil)
-    }
-    
-    // Handle Incoming pushes
-    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType, completion: @escaping () -> Void) {
-        // MARK: Handling incoming call
-        SendBirdCall.pushRegistry(registry, didReceiveIncomingPushWith: payload, for: type) { uuid in
-            guard uuid != nil else {
-                let update = CXCallUpdate()
-                update.remoteHandle = CXHandle(type: .generic, value: "invalid")
-                let randomUUID = UUID()
-
-                CXCallManager.shared.reportIncomingCall(with: randomUUID, update: update) { error in
-                    CXCallManager.shared.endCall(for: randomUUID, endedAt: Date(), reason: .acceptFailed)
-                }
-                completion()
-                return
-            }
-
-            completion()
-        }
-    }
-}
-
-// MARK: - Remote Notification
-extension AppDelegate {
-    func remoteNotificationsRegistration(_ application: UIApplication) {
-        application.registerForRemoteNotifications()
-        
-        let center = UNUserNotificationCenter.current()
-        center.requestAuthorization(options: [.alert, .badge, .sound]) { success, error in
-            guard error == nil else {
-                print("Error while requesting permission for notifications.")
-                return
-            }
-            
-            // If success is true, the permission is given and notifications will be delivered.
-        }
-    }
-    
-    func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
-        // Check your app's configurations for APNs.
-    }
-    
-    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
-        UserDefaults.standard.remotePushToken = deviceToken
-        SendBirdCall.registerRemotePush(token: deviceToken) { error in
-            print("Remote Notifications Device token is \(deviceToken.toHexString())")
-        }
-    }
-    
-    func application(_ application: UIApplication, didReceiveRemoteNotification userInfo: [AnyHashable : Any], fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
-        if userInfo.keys.contains("sendbird_call") { // This is not necessary; SendBirdCall will not process and invoke the completionHandler for payloads that don't contain the key "sendbird_calls".
-            SendBirdCall.application(application, didReceiveRemoteNotification: userInfo)
-        }
-    }
-}
-
